+++ src/common/store.tsx
@@ -1,0 +1,16 @@
+import { applyMiddleware, createStore } from 'redux';
+import { rootReducer, rootEpic } from '../reducers';
+import { createEpicMiddleware } from 'redux-observable';
+
+export const createReduxStore = (initialState = {}) => {
+    const epicMiddleware = createEpicMiddleware();
+    const middleware: any[] = [epicMiddleware];
+
+    const store = createStore(
+        rootReducer,
+        initialState,
+        applyMiddleware(...middleware)
+    );
+    epicMiddleware.run(rootEpic);
+    return store
+};
