+++ src/components/weekCalendar/WeeklyCalenldar.tsx
@@ -1,0 +1,63 @@
+import * as React from 'react';
+import * as moment from 'moment';
+import styled from 'styled-components';
+import { Task } from '../../models/Task';
+import { DayColumn } from './DayColumn';
+import { range } from 'lodash';
+import { DayOfWeekHeader } from '../DayOfWeekHeader';
+import { getDaysInWeek, isMatchDate } from '../../utils/calendarUtil';
+
+const Container = styled.div`
+    text-align: center;
+    padding-left: 50px;
+`
+const Context = styled.div`
+    position: relative;
+    display: flex;
+`
+const DateHeader = styled.div`
+    display: flex;
+    flex-flow: row;
+    font-size: 3em;
+    & > * {
+        flex: 1;
+    }
+`
+const TimeContainer = styled.div`
+    position: absolute;
+    left: -40px;
+    top: 30px;
+`
+
+interface OwnProps {
+    currentMoment: moment.Moment
+    tasks: Task[]
+}
+
+type Props = OwnProps & React.HTMLAttributes<HTMLDivElement>;
+
+export const WeeklyCalendar: React.FC<Props> = (props: Props) => {
+    const { currentMoment, tasks, ...divProps } = props;
+    const hours = currentMoment.clone().endOf('day').hour();
+    const daysInWeek = getDaysInWeek(currentMoment);
+    const restWeek = currentMoment.clone().startOf('week');
+    const dates = range(0, daysInWeek).map((i) => restWeek.clone().add(i, 'd'));
+    const columns = dates.map(m => {
+        const filteredTasks = tasks.filter(task => isMatchDate(moment(task.date), m));
+        return (<DayColumn key={m.toString()} dateMoment={m} tasks={filteredTasks}/>);
+    });
+    return (
+        <Container {...divProps}>
+            <DayOfWeekHeader />
+            <DateHeader>{dates.map(m => <div key={m.date()}>{m.date()}</div>)}</DateHeader>
+            <Context>
+                <TimeContainer>
+                    {range(1, hours + 1).map(time => <div key={time} style={{height: '40px'}}>{time} 시</div>)}
+                </TimeContainer>
+                {columns}
+            </Context>
+
+
+        </Container>
+    )
+}
