+++ src/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import {App} from './containers/App';
-import {createReduxStore} from './libs/redux/store';
+import {createReduxStore} from './common/store';
 import './index.css';
 import registerServiceWorker from './registerServiceWorker';
