+++ src/reducers/task/index.ts
@@ -1,0 +1,15 @@
+import { ACTION_HANDLERS } from './handler';
+import { Task } from '../../models/Task';
+import { handleActions } from '../../common/immutableReducer';
+
+export class TaskStoreModel {
+    constructor(
+        public tasks: Task[] = [],
+        public isTaskLoading: boolean = false,
+        public isTaskLoaded: boolean = false,
+        public isTaskUpdating: boolean = false,
+        public isTaskUpdated: boolean = false
+    ) { }
+}
+
+export const reducer = handleActions<TaskStoreModel>(new TaskStoreModel(), ACTION_HANDLERS);
